better 64-bit support

+++ src/common/lzma/lzma.h
@@ -47,7 +47,7 @@
 unsigned char *LZMA_Compress( 
 unsigned char	*pInput,
 unsigned int	inputSize, 
-unsigned int	*pOutputSize, 
+size_t			*pOutputSize, 
 unsigned int	dictionarySize = LZMA_DEFAULT_DICTIONARY );
 
 //-----------------------------------------------------------------------------
